src/requirements.txt
pip==23.2.1
wheel==0.41.1
setuptools==68.0.0
Django==4.2.4
asgiref==3.7.2
sqlparse==0.4.4
requests==2.31.0
urllib3==2.0.4
mock==5.1.0 # for testing
testtools~=2.6.0 # for testing
fixtures==4.1.0 # for testing
qrcode ~= 7.4.2
django-environ==0.10.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0